Added some info and pics to README
